# rkt/image: render images on fetch