--- a/tests/rkt_image_dependencies_test.go
+++ b/tests/rkt_image_dependencies_test.go
@@ -18,9 +18,11 @@ import (
 	"fmt"
 	"io/ioutil"
 	"os"
+	"path/filepath"
 	"strings"
 	"testing"
 
+	"github.com/coreos/rkt/common"
 	"github.com/coreos/rkt/tests/testutils"
 	taas "github.com/coreos/rkt/tests/testutils/aci-server"
 )
@@ -62,20 +64,22 @@ const (
 `
 )
 
-// TestImageDependencies generates ACIs with a complex dependency tree and
-// fetches them via the discovery mechanism. Some dependencies are already
-// cached in the CAS, and some dependencies are fetched via the discovery
-// mechanism. This is to reproduce the scenario in explained in:
-// https://github.com/coreos/rkt/issues/1752#issue-117121841
-func TestImageDependencies(t *testing.T) {
+var topImage string = "localhost/image-a"
+
+type testImage struct {
+	shortName string
+	imageName string
+	deps      string
+	version   string
+	prefetch  bool
+
+	manifest string
+	fileName string
+}
+
+func generateComplexDependencyTree(t *testing.T, ctx *testutils.RktRunCtx) (map[string]string, []testImage) {
 	tmpDir := createTempDirOrPanic("rkt-TestImageDeps-")
 	defer os.RemoveAll(tmpDir)
-
-	ctx := testutils.NewRktRunCtx()
-	defer ctx.Cleanup()
-
-	server := runServer(t, taas.GetDefaultServerSetup())
-	defer server.Close()
 
 	baseImage := getInspectImagePath()
 	_ = importImageAndFetchHash(t, ctx, "", baseImage)
@@ -96,17 +100,7 @@ func TestImageDependencies(t *testing.T) {
 	// D->B
 	// D->E
 
-	topImage := "localhost/image-a"
-	imageList := []struct {
-		shortName string
-		imageName string
-		deps      string
-		version   string
-		prefetch  bool
-
-		manifest string
-		fileName string
-	}{
+	imageList := []testImage{
 		{
 			shortName: "a",
 			imageName: topImage,
@@ -161,8 +155,27 @@ func TestImageDependencies(t *testing.T) {
 
 		baseName := "image-" + img.shortName + ".aci"
 		img.fileName = patchACI(emptyImage, baseName, "--manifest", tmpManifest.Name())
-		defer os.Remove(img.fileName)
 		fileSet[baseName] = img.fileName
+	}
+
+	return fileSet, imageList
+}
+
+// TestImageDependencies generates ACIs with a complex dependency tree and
+// fetches them via the discovery mechanism. Some dependencies are already
+// cached in the CAS, and some dependencies are fetched via the discovery
+// mechanism. This is to reproduce the scenario in explained in:
+// https://github.com/coreos/rkt/issues/1752#issue-117121841
+func TestImageDependencies(t *testing.T) {
+	ctx := testutils.NewRktRunCtx()
+	defer ctx.Cleanup()
+
+	server := runServer(t, taas.GetDefaultServerSetup())
+	defer server.Close()
+
+	fileSet, imageList := generateComplexDependencyTree(t, ctx)
+	for _, img := range imageList {
+		defer os.Remove(img.fileName)
 	}
 
 	if err := server.UpdateFileSet(fileSet); err != nil {
@@ -194,6 +207,52 @@ func TestImageDependencies(t *testing.T) {
 		if err := expectWithOutput(child, expected); err != nil {
 			t.Fatalf("Expected %q but not found: %v", expected, err)
 		}
+	}
+
+	waitOrFail(t, child, 0)
+}
+
+func TestRenderOnFetch(t *testing.T) {
+	// If overlayfs is not supported, we don't render images on fetch
+	if !common.SupportsOverlay() {
+		t.Skip("Overlay fs not supported.")
+	}
+	ctx := testutils.NewRktRunCtx()
+	defer ctx.Cleanup()
+
+	server := runServer(t, taas.GetDefaultServerSetup())
+	defer server.Close()
+
+	fileSet, imageList := generateComplexDependencyTree(t, ctx)
+	for _, img := range imageList {
+		defer os.Remove(img.fileName)
+	}
+
+	if err := server.UpdateFileSet(fileSet); err != nil {
+		t.Fatalf("Failed to populate a file list in test aci server: %v", err)
+	}
+
+	for i := len(imageList) - 1; i >= 0; i-- {
+		img := imageList[i]
+		if img.prefetch {
+			t.Logf("Importing image %q: %q", img.imageName, img.fileName)
+			testImageShortHash := importImageAndFetchHash(t, ctx, "", img.fileName)
+			t.Logf("Imported image %q: %s", img.imageName, testImageShortHash)
+		}
+	}
+
+	fetchCmd := fmt.Sprintf("%s --debug --insecure-options=image,tls fetch %s", ctx.Cmd(), topImage)
+	child := spawnOrFail(t, fetchCmd)
+
+	treeStoreDir := filepath.Join(ctx.DataDir(), "cas", "tree")
+	trees, err := ioutil.ReadDir(treeStoreDir)
+	if err != nil {
+		panic(fmt.Sprintf("Cannot read tree store dir: %v", err))
+	}
+
+	// We expect 2 trees: stage1 and the image
+	if len(trees) != 2 {
+		t.Fatalf("Expected 2 trees but found %d", len(trees))
 	}
 
 	waitOrFail(t, child, 0)