rkt was delaying rendering images to the tree store until they were
about to run for the first time which caused that first run to be slow
for big images.

This commit renders the image to the tree store at fetch time.

If the system doesn't support overlayfs, we don't render it because we
don't use the tree store when overlayfs is not used.

Startup time benchmark:

rkt run a Docker nginx image, pre-fetched. Size = 188MiB when unextracted
The insecure option `ondisk` was used.

command | time (rkt master) | time (rkt master + https://github.com/coreos/rkt/pull/2398)
--------|-------------------|---------------------------
1st time run | real 0m4.964s, user 0m3.570s, sys 0m0.813s | real 0m0.333s, user 0m0.097s, sys 0m0.050s
2nd time run | real 0m0.350s, user 0m0.063s, sys 0m0.077s | real 0m0.343s, user 0m0.073s, sys 0m0.070s

Fixes #2388