@jonboulle 2016-04-08
We should probably also have a specific issue  about parallelising
rendering

Iago López Galeiras <notifications@github.com> schrieb am Fr., 8. Apr. 2016
15:48:

> rkt was delaying rendering images to the tree store until they were
> about to run for the first time which caused that first run to be slow
> for big images.
>
> This commit renders the image to the tree store at fetch time.
>
> If the system doesn't support overlayfs, we don't render it because we
> don't use the tree store when overlayfs is not used.
>
> Fixes #2388 <https://github.com/coreos/rkt/issues/2388>
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/coreos/rkt/pull/2398
> Commit Summary
>
>    - rkt/image: render images on fetch
>
> File Changes
>
>    - *M* rkt/image/fetcher.go
>    <https://github.com/coreos/rkt/pull/2398/files#diff-0> (22)
>
> Patch Links:
>
>    - https://github.com/coreos/rkt/pull/2398.patch
>    - https://github.com/coreos/rkt/pull/2398.diff
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly or view it on GitHub
> <https://github.com/coreos/rkt/pull/2398>
>