@iaguis 2016-04-08
> As I remember stage1 rendering is special since the treestore is always rendered also if not using overlayfs and then the stage1 is copied from it.

That's correct.