@iaguis 2016-04-08
> But I'm not sure if it's worth to handle this case.

In that case, we will render it when running the first ACI and then copy it. The next runs will only copy it. I think it's an OK behavior.