@sgotti 2016-04-08
> In that case, we will render it when running the first ACI and then copy it. The next runs will only copy it. I think it's an OK behavior.

Ok for me too. I don't see a great speed gain since the stage1 won't change often and will be copied anyway at every run.