@iaguis 2016-04-11
Added some benchmarks.