@sgotti 2016-04-14
I have a doubt that this won't work when a fetched images has some dependencies since the dependencies are fetched with multiple calls to FetchSingleImage. I have to test this case.
Additionally this will render a treestore for every image dependency.

Probably the rendering should be inserted at the end of Fetcher.FetchImage

Perhaps some tests are needed to catch this possible issue.