@iaguis 2016-04-14
Hm, you're right...

On Thu, Apr 14, 2016, 13:14 Simone Gotti <notifications@github.com> wrote:

> I have a doubt that this won't work when a fetched images has some
> dependencies since the dependencies are fetched with multiple calls to
> FetchSingleImage. I have to test this case.
> Additionally this will render a treestore for every image dependency.
>
> Probably the rendering should be inserted at the end of Fetcher.FetchImage
>
> Perhaps some tests are needed to catch this possible issue.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly or view it on GitHub
> <https://github.com/coreos/rkt/pull/2398#issuecomment-209885664>
>