@iaguis 2016-04-14
Changed to render on `FetchImage()`. I'll write some tests.