@yifan-gu 2016-04-14
Sorry, I am not able to review this today. ( I am digging into the e2e tests failing).