@yifan-gu 2016-04-15
As for stage1 rendering,  #2364 might improve it.