@iaguis 2016-04-19
Added a functional test.